refactor(LARA-215): extract FocusTile from focus-host-slot host (no behavior change)
